+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Severity filter for detectors (optimization): Issue [#358](https://github.com/nowarp/misti/issues/358)
+
 ## [0.8.0] - 2025-04-08
 
 ### Added