+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `PreferredStdlibApi` is enabled by default; `PreferAugmentedAssign` is disabled
 - Prevent multiple `Tool`s with plain output format: Issue [#363](https://github.com/nowarp/misti/issues/363)
 - Forbid executing `Dump{Cfg,CallGraph,Imports}` without input file: Issue [#363](https://github.com/nowarp/misti/issues/363)
+- `package.json`: Update Misti version on every `yarn build`
 
 ### Fixed
 - Severity filter for detectors (optimization): Issue [#358](https://github.com/nowarp/misti/issues/358)