+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SuboptimalSend`: Incorrect `message`->`cashback` suggestion: Issue [#366](https://github.com/nowarp/misti/issues/366)
 - `Logger`: Show context in JSON output: Issue [#370](https://github.com/nowarp/misti/issues/370)
 - `Logger`: Race condition for context in asynchronous execution
+- Unused `ResultError` and exit code `2` is never set: Issue [#371](https://github.com/nowarp/misti/issues/371)
 
 ## [0.8.0] - 2025-04-08