+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.1] - 2025-04-19
+
 ### Added
 - Tact 1.6.6+ support: PR [#368](https://github.com/nowarp/misti/pull/368)
 - `Driver`: Save detectors execution time in debug log